backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-cors==4.0.0
authlib==1.2.1
requests==2.31.0
werkzeug==2.3.7
flask-migrate==4.0.5
PyJWT==2.8.0
